fixup! Update schema with changes
